Move over process code to engine & fix headers.
