+++ include/engine/esprite.h
