+++ include/functions.us.h
@@ -44,9 +44,6 @@ void func_800A14F0(Vec3f*, Vec3f*, Vec3f*);
 void func_800410E8(u8);
 s32 func_80041624(u8);
 void func_8005DDDC(void);
-Process* CreateChildProcess(process_func func, u16 priority, s32 stack_size, s32 extra_data_size, Process* parent);
-Process* CreateProcess(process_func func, u16 priority, s32 stack_size, s32 extra_data_size);
-void SetProcessDestructor(Process* process, process_func destructor);
 void func_80025F60(s16, s32);
 void func_80041048(u8, s32);
 void func_800413B0(u8);
@@ -193,8 +190,6 @@ s16 GetCurrentPlayerIndex(void);
 void func_80045E6C(s32);
 s32 func_80045D84(s32, s32, s32);
 s16 RunDecisionTree(DecisionTreeNonLeafNode*);
-void LinkChildProcess(Process* process, Process* child);
-void WaitForChildProcess(void);
 void func_8004CD84(Vec3f*);
 Process* func_8004D1EC(Vec3f*, Vec3f*, Vec3f*, s32);
 Object* func_8003E320(Object*);
@@ -258,11 +253,8 @@ void func_8004F2EC(void);
 void func_8004F5F0(void);
 void func_80067704(s16);
 void func_80067384(s16, s16, u16);
-void KillChildProcess(Process* process);
-s32 KillProcess(Process* process);
 s32 setjmp(jmp_buf* jump_buf);
 s32 longjmp(jmp_buf* jump_buf, s32 val);
-void UnlinkChildProcess(Process* process);
 s32 GetMemoryAllocSize(s32 value);
 void* MakeHeap(void* ptr, u32 size);
 void Free(void* ptr);
@@ -274,7 +266,6 @@ s32 GetMemoryAllocSize(s32 value);
 s32 EndProcess(Process*);
 void func_8003E1BC(Object*, u8);
 void SleepVProcess(void);
-Process* GetCurrentProcess(void);
 void func_8003E8B8(Object* ptr, s16 a, s16 b, s16 c, u16 d);
 void DecodeFile(void* src, void* dest, s32 len, s32 decode_type);
 void* MallocTemp(u32 size);
@@ -296,7 +287,6 @@ void func_800421E0(void);
 unkObjectStruct* func_8004E3E0(s32, Vec3f*, s32, void*); //arg2 should be Vec3f*
 void func_8004F4D4(void*, s32, s32);
 void func_80055960(s32, s32);
-void SleepProcess(s32);
 void func_8005D96C(unkObjectStruct*, f32, f32, f32);
 void func_8005D97C(unkObjectStruct*, f32, f32, f32);
 void InitCameras(s32);
@@ -358,7 +348,6 @@ void func_8004F284(void);
 void func_8004F28C(s32, s16);
 void func_800546B4(s32, s32);
 void func_8005E3A8(void);
-void SleepVProcess(void);
 void func_800568A4(void);
 void ClearBoardFeatureFlag(s32 flag);
 void SetBoardFeatureFlag(s32 flag);
