+++ src/ovl_36_DKsJungleAdventure/2418D0.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "process.h"
 #include "dkJungleAdventure.h"
 
 // bss
