+++ src/ovl_3D_EternalStar/254480.c
@@ -1,4 +1,3 @@
-#include "common.h"
 #include "eternalStar.h"
