+++ src/ovl_3D_EternalStar/eternalStar.h
@@ -1,4 +1,5 @@
 #include "common.h"
+#include "process.h"
 
 //start of rodata 800F8B80
 //bss 800F9160 - 800F91B0
